requirements.txt
jupyter==1.0.0
pandas==0.18.1
matplotlib==1.5.2
seaborn==0.7.1
tqdm==4.8.4
lifelines==0.9.2
statsmodels==0.6.1
python-igraph==0.7.1post6
